+++ dealer-server/src/main/java/com/autodealer/crm/service/impl/ClueServiceImpl.java
@@ -3,6 +3,7 @@ package com.autodealer.crm.service.impl;
 import com.alibaba.excel.EasyExcel;
 import com.autodealer.crm.config.converter.ClueExcelConverter;
 import com.autodealer.crm.config.listener.UploadDataListener;
+import com.autodealer.crm.config.security.CurrentUserProvider;
 import com.autodealer.crm.constant.Constants;
 import com.autodealer.crm.mapper.TClueMapper;
 import com.autodealer.crm.mapper.TClueRemarkMapper;
@@ -11,7 +12,6 @@ import com.autodealer.crm.query.BaseQuery;
 import com.autodealer.crm.query.ClueQuery;
 import com.autodealer.crm.result.ClueExcel;
 import com.autodealer.crm.service.ClueService;
-import com.autodealer.crm.util.JWTUtils;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import jakarta.annotation.Resource;
@@ -34,6 +34,9 @@ public class ClueServiceImpl implements ClueService {
 
     @Resource
     private ClueExcelConverter clueExcelConverter;
+
+    @Resource
+    private CurrentUserProvider currentUserProvider;
 
     @Override
     public PageInfo<TClue> getClueByPage(Integer current, Integer pageSize) {
@@ -59,9 +62,10 @@ public class ClueServiceImpl implements ClueService {
     }
 
     @Override
-    public void importExcel(InputStream inputStream, String token) {
+    public void importExcel(InputStream inputStream) {
         //链式编程，3个参数, 第一个参数是要读取的Excel文件，第二个参数是Excel模板类，第三个参数是文件读取的监听器
-        EasyExcel.read(inputStream, ClueExcel.class, new UploadDataListener(tClueMapper, token, clueExcelConverter))
+        EasyExcel.read(inputStream, ClueExcel.class,
+                        new UploadDataListener(tClueMapper, currentUserProvider.getCurrentUserId(), clueExcelConverter))
                 .sheet()
                 .doRead();
     }
@@ -83,11 +87,10 @@ public class ClueServiceImpl implements ClueService {
             //Spring框架有个工具类BeanUtils可以进行对象的复制,复制的条件要求是：两个对象的字段名要相同，字段的类型也相同，这样才可以复制
             BeanUtils.copyProperties(clueQuery, tClue);
 
-            //解析jwt得到userId
-            Integer loginUserId = JWTUtils.parseUserFromJWT(clueQuery.getToken()).getId();
-
+            Integer operatorId = currentUserProvider.getCurrentUserId();
+            tClue.setOwnerId(operatorId);
             tClue.setCreateTime(new Date()); //创建时间
-            tClue.setCreateBy(loginUserId); //创建人id
+            tClue.setCreateBy(operatorId); //创建人id
 
             return tClueMapper.insertSelective(tClue);
         } else {
@@ -97,34 +100,29 @@ public class ClueServiceImpl implements ClueService {
 
     @Override
     public TClue getClueById(Integer id) {
-        return tClueMapper.selectDetailById(id);
+        return tClueMapper.selectDetailById(id, currentUserProvider.getDataScopeUserId());
     }
 
     @Transactional(rollbackFor = Exception.class)
     @Override
     public int updateClue(ClueQuery clueQuery) {
         // 先查询原记录，获取原手机号
-        TClue originalClue = tClueMapper.selectByPrimaryKey(clueQuery.getId());
-        if (originalClue == null) {
-            throw new RuntimeException("线索记录不存在");
-        }
+        TClue originalClue = requireAccessibleClue(clueQuery.getId());
 
         TClue tClue = new TClue();
 
         //把前端提交过来的参数数据对象ClueQuery复制到TClue对象中
         //Spring框架有个工具类BeanUtils可以进行对象的复制,复制的条件要求是：两个对象的字段名要相同，字段的类型也相同，这样才可以复制
         BeanUtils.copyProperties(clueQuery, tClue);
+        tClue.setOwnerId(originalClue.getOwnerId());
 
         // 如果传入的手机号与原记录不同，忽略手机号字段
         if (clueQuery.getPhone() != null && !clueQuery.getPhone().equals(originalClue.getPhone())) {
             tClue.setPhone(null); // 设置为null，让MyBatis的updateByPrimaryKeySelective跳过该字段
         }
 
-        //解析jwt得到userId
-        Integer loginUserId = JWTUtils.parseUserFromJWT(clueQuery.getToken()).getId();
-
         tClue.setEditTime(new Date()); //编辑时间
-        tClue.setEditBy(loginUserId); //编辑人id
+        tClue.setEditBy(currentUserProvider.getCurrentUserId()); //编辑人id
 
         return tClueMapper.updateByPrimaryKeySelective(tClue);
     }    
@@ -135,6 +133,7 @@ public class ClueServiceImpl implements ClueService {
         if (id == null) {
             return 0;
         }
+        requireAccessibleClue(id);
         // 先删除关联的线索备注
         tClueRemarkMapper.deleteByClueId(id);
         // 再删除线索
@@ -147,11 +146,22 @@ public class ClueServiceImpl implements ClueService {
         if (ids == null || ids.size() == 0) {
             return 0;
         }
+        List<Integer> distinctIds = ids.stream().distinct().sorted().toList();
+        distinctIds.forEach(this::requireAccessibleClue);
         // 先删除关联的线索备注
-        for (Integer id : ids) {
+        for (Integer id : distinctIds) {
             tClueRemarkMapper.deleteByClueId(id);
         }
         // 再删除线索
-        return tClueMapper.batchDeleteByIds(ids);
+        return tClueMapper.batchDeleteByIds(distinctIds);
+    }
+
+    private TClue requireAccessibleClue(Integer id) {
+        TClue clue = tClueMapper.selectScopedByPrimaryKey(
+                id, currentUserProvider.getDataScopeUserId());
+        if (clue == null) {
+            throw new RuntimeException("线索不存在或无权访问");
+        }
+        return clue;
     }
 }
